Add unit tests (#4)
* Require fields when decoding JSON

Add `validate:"required"` tag to struct fields that are required. The `DecodeJson` function now respects these required fields, and returns an error if they are not present in the request body.

* Pass in graph to handler functions

Previously, the `handlers` package instantiated its own graph. However, this reduced flexibility and caused unit tests to fail because it was instantiated in the `init` function, and not reset between unit tests. Instead, it is now passed into the handler functions.

* Add new error types

Comparing error types is cleaner than comparing error messages

* Add unit tests